refactor[#47] : /write 페이지 리팩토링
